requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
pydantic
requests
pytest
httpx
python-multipart
openai
python-dotenv
google-generativeai
authlib
python-jose[cryptography]
